refactor: implement loot composable
