+++ src/packages/frontend/src/components/Team.vue
@@ -7,35 +7,19 @@ import TeamName from './TeamName.vue'
 import ProfitEach from './ProfitEach.vue'
 import { computed } from 'vue'
 import {
-  calcualteTotalLootValue,
-  filterLoot,
-  groupLoot,
-} from '../utils/loot/loot.helpers.ts'
-import { useConfigStore } from '../stores/configStore.ts'
-import { useLootDataStore } from '../stores/lootDataStore.ts'
-import {
   getMembersWithWaste,
   getTotalWaste,
 } from '../utils/waste/waste.utils.ts'
+import { useLoot } from '../composables/useLoot.ts'
 
-const configStore = useConfigStore()
-const lootDataStore = useLootDataStore()
+const { totalLootValue } = useLoot()
 
 const { team } = defineProps<{ team: Team }>()
 const emit = defineEmits(['updateName'])
 
-const loot = computed(() => {
-  const lootFiltered = filterLoot(
-    lootDataStore.lootDataParsed.loot,
-    configStore.config,
-  )
-  return groupLoot(lootFiltered)
-})
-
 const profitEach = computed(() => {
-  const totalLootValue = calcualteTotalLootValue(loot.value)
   const totalWasteAmount = getTotalWaste(team.wastes)
-  const totalProfit = totalLootValue - totalWasteAmount
+  const totalProfit = totalLootValue.value - totalWasteAmount
   const membersWithWasteAmount = getMembersWithWaste(team.wastes).length
 
   if (membersWithWasteAmount === 0) return null
